requirements.txt
affine==2.1.0
aiohttp==2.3.1
async-timeout==2.0.0
attrs==17.2.0
chardet==3.0.4
click==6.7
coverage==4.4.1
mercantile==0.11.0
multidict==3.3.0
numpy==1.13.3
olefile==0.44
Pillow==4.3.0
py==1.4.34
scipy==0.19.1
yarl==0.13.0
